First attempt at adding platform to BaseDependency
This is complicated because of the way the pyproject.toml source
parsing is structured. Breaking out into platforms only happens at the
end when the markers are evaluated, so for most of the time
the platform has not yet been determined. It wouldn't be that difficult
to fix, but it doesn't seem worth it at the moment, but I'm leaving this
here in case I change my mind.